Add tracing support
